requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-genai>=1.51.0
python-dotenv
python-multipart
sqlalchemy
asyncpg
openai>=1.120.0
tenacity